Pass only necessary args to draw functions
Instead of passing a full json object I only use once, just pass the
final dimension width, since that's all the function is using. I'm not
sure if my future plans for the web version involve a json object, so
dropping this dependency will make it easier to decide on options later.