Update 9900CC oracle image
This image was failing the test when run on a different OS. When
comparing all of the images, we determined that the new version is
actually more consistent with the other oracle images. The only minor
difference is the amount of rounding on the rectangles' corners. Our
best explanation for why only this test image was failing is that the size of
the source and output image changes the stroke width which causes a
slight rounding difference somewhere in the Fabric/Canvas/Cairo stack.