Line: fix anti-aliased borders
For vertical and horizontal lines, edges can easily end up on pixel
borders.  Anti-aliasing those looks pretty bad, but we can fix it just
like we did with rectangles..

Also, fix the occasional disappearing issue we saw when a line was
sized through one of the axes. It happened when scale was briefly set
to 0. Never again!