Shapes: add mixin.js
This allows mixins to override existing methods, but still call them
with callParent().

Previously, a method from a mixin would just override any other mixin
method with the same name and `callSuper()` woulc call the super-class
method, not function it overrode.

Mostly, this was added so the _set in limit_size.js and the _set in
two_point_interactivity.js dont clobber eachother.

The only weird thing is that a method in a Class definition needs to
use `callSuper(name, params...)` while a method in a mixin needs to
use `callParent(params)`.