Lines: enforce min/max correctly
The previous method Had several issues:
* Could never make a vertical or horizontal line because of minW/minH
* Min/Max were enforced in a way that limited the line to a box
  instead of a circle

Now we enforce the limits in a sepcial way if min/max refers to a
diagonal. Particularly, we test if the new scale would push the whole
shape beyond its limits. If so, we alter both scaleX and scaleY so the
overall length is equal to one of the limits (and ignore the original
_set() operation).