Lines: include endpoints in toObject
If a line has been sized after drawing it, the endpoints at obj.x1,
obj.y1 are not the actual endpoints and are only used for
detecting the line orientation.

Previously you'd end up with a line of the original
drawn size, now you get one just as you see it.