Sub-pixel alignment: improve
After the new fabric version and changes to drawing methods,
sub-pixel alignment and calculation of border-widths was off.
I tweaked it till it played nice with the existing test
images and added it to line-based shapes as well.

Alter W and H so lines are rendered perfectly vertical or horizontal.