drag-to-create: use screen offsets for marker size
There was a bug during marker creation. If you drug the mouse outside
the canvas while draring a new marker, the marker would jump around
seemingly randomly.

We are using the .offsetX and .layerX properties of mouse events.
These give the coordinates relative to the target's upper left coord.
This works fine for the mousedown point cause that's guaranteed to be
in the canvas. During mousemove, the target can be *some other
element* and thus the coordinates are wrong.

Solution
========
We use the mousedown point as we have been; But when computing the
mouse move position, we compare .pageX/Y (absolute page coords) of
each mouse event to get an offset from the original mousedown event.
This way we're able to find the coords relative to the canvas for
both mousedown and mousemove.