Lines: fix incrementSize() with negative values
Previously, it was enforcing the minimums on a per-axis basis. This
doesn't work for close-to-vertical or close-to-horizontal lines.  They
hit the minimum in one dimension while still decresing in the other.
This leads to a line that changes orientation as it gets smaller.

Now we actually use the diagonal checking function and just bail out
(not making any changes) if it would have pushed us over the limit.