Drag to Create: fix mouse positions for FF 41
The previous method using offsetX and layerX was not cross-browser
and relied on undocumented behavior. The newer versions of FF change
the behavior and markup would show in a crazy wrong place.

Now, we use the canvas functions to get the mouse position.

pageX/Y are standard and cross-browser, so we're safe with those.