Restructure test file
Python doesn't hoist. But the idomatic pattern to work around that is
defining our `main` function at the top of the file, and using this
magic:
```
if __name__ == "__main__":
   main()
```
to automatically run the `main` entrypoint function if this file is
executed. We can also then use this file as a library as well,
`import`ing, but not immediately kicking off the `main` function.