Unify structure of both copies of `cleanJSON`
Each of the two main entrypoint modules define a `cleanJSON` function.
It appears to be working to consistently cast integer fields from
strings. But we have two subtly different implementations. Mirroring the
structure shows that `ImageMarkupCall` does not cast `strokeWidth` to an
integer value. Otherwise, these functions are (now) identical.