Make `size` variable names unique in Call
We were getting the complaint here that `size` was being redeclared.
Lets make the variable name `size` in this file unique to usage/context.

The assignments here were happening in the same scope, in the same loop.
But the good news it is unlikely to cause any trouble, as the assignment
is _unconditional_, in the specific switch `case`.

Detected by Eslint as:
```
   87:15  error  'size' is already defined        no-redeclare
  134:15  error  'size' is already defined        no-redeclare
```