ImageMarkupCall: Use Yargs features
Unfortunately, `yargs` doesn't do the right thing with command names
that look like options. This switches us to using `json` and `markup`
as the command names. Grepping suggests that we'll only need to change
one place in the main repo and the `run-tests.py` script here to
accommodate the new syntax. Switching the syntax lets us take
advantage of Yargs features like only allowing specific options for
specific commands. We're able to eliminate all of our validation as a
result.

One odd bit: I had to use the `check()` method rather than the
`demandCommand()` method (which seems like it's the correct way to
enforce that exactly one command at a time is used), because
`demandCommand()` wasn't failing runs with both commands (despite the
fact that I'd set the `max` number).

Co-authored-by: Daryl Metzler <daryl@ifixit.com>