Shim back flag form commands
Yargs can't correctly parse flag form commands. `--` will always prefix
an "option", and using that prefix on commands, even aliases, does not
work.

We can use the "default command" pattern in yargs:
See:
To allow passing in `--json` and `--markup` as "options", and then shim
them to "commands". That's not the most elegant solution, but it is
workable, at least until we deprecate usages of the flag form arguments.