Simplify iterating over instructions
That for loop is pretty tough to read. It starts at instruction zero,
but skips that via the pre-increment (`++i`) to walk past the first
empty split. Then, we only traverse to `i < instructions.length` in
order to skip the last element of the list as well.

That's a lot of tough to miss trickery, and in fact, not relevant. The
condtional guard on the first line of the loop skips empty instructions
anyhow. Lets iterate the entire list.

Note that `continue` is not allowed in a `forEach`. We'll fix that next.