Call Call directly
We're using an odd bash wrapper to exec ImageMarkupCall from the python
test. In order to avoid needing to install `bash` on alpine (we shouldnt
need it) lets exec `node` directly. This skips adding the system node
modules path to `NODE_PATH`, but that's something we're trying to get
done anyhow.