Hacks to get positioning correct
Wat. The rectangle and line shapes seem to be off position by a half
borderwidth. With these bandaids in place, the visual diff on the test
images is basically nothing. Only zooming way in shows any real
difference, and it appears to be mostly anti-aliasing related. Lets give
this a try for the short term to get our library up and working.