Avoid node-deps in Builder
Specifically, we want to import our Builder in the browser context.
Currently, this fails hard and fast at our `require` of `path` in the
Builder.

In this commit: https://github.com/iFixit/node-markup/commit/ee1c72a3c4b39fbad87d2bba88e51c3ddc4162d4
during the Fabric upgrade, we switched to passing a full `file://` URL
in the options to `Fabric.Image.fromObject`. This was to handle a change
to the Fabric internals where we now require an URL.

We can use `path.resolve` to build our `file://` URL, but `path` is a
Node specific dependency, and breaks at webpack (can't resolve that
module without a polyfill.)

We can avoid the trouble by pre-resolving this path/URL in the outer
`ImageMarkupCall` wrapper. Since that file is the node-specific
entrypoint, it is safe for us to require and use `path` here.

This keeps the node specific dependency inside the node entrypoint, and
should avoid needing to polyfill `path` on the browser side.